Prepare objects for plotting or plot objects

data_plot() attempts to extract and tranform an object to be further plotted, while plot() tries to visualize results of functions from different packages of the easystats-project.

data_plot(x, data = NULL, ...)

# S3 method for see_bayesfactor_models
plot(x, n_pies = c("one", "many"),
  value = c("none", "BF", "probability"), log = FALSE,
  prior_odds = NULL, ...)

# S3 method for see_bayesfactor_savagedickey
plot(x, point_size = 2, ...)

# S3 method for see_check_distribution
plot(x, point_size = 2, panel = TRUE,
  ...)

# S3 method for see_check_normality
plot(x, type = c("density", "qq", "pp"),
  data = NULL, ...)

# S3 method for see_check_outliers
plot(x, text_size = 3.5, ...)

# S3 method for see_equivalence_test
plot(x, rope_color = "#0171D3",
  rope_alpha = 0.2, show_intercept = FALSE, ...)

# S3 method for see_estimate_density
plot(x, stack = TRUE,
  show_intercept = FALSE, ...)

# S3 method for see_rope
plot(x, data = NULL, rope_alpha = 0.5,
  rope_color = "cadetblue", show_intercept = FALSE, ...)

Arguments

x
An object.

data
The original data used to create this object. Can be a statistical model or such.

...
Arguments passed to or from other methods.

n_pies
Number of pies.

value
What value to display.

log
Show log-transformed Bayes factors.

prior_odds
optional vector of prior odds for the models. See BayesFactor::priorOdds. As the size of the pizza slices corresponds to posterior probability (which is a function of prior probability and the BF), custom prior_odds will change the slices' size.

point_size
Size of point-geoms.

panel
Logical, if TRUE, plots are arranged as panels; else, single plots are returned.

type
Character vector, indicating the type of plot (for check_normality).

text_size
Size of text labels.

rope_color
Color of ROPE ribbon.

rope_alpha
Transparency level of ROPE ribbon.

show_intercept
Logical, if TRUE, the intercept-parameter is included in the plot. By default, it is hidden because in many cases the intercept-parameter has a posterior distribution on a very different location, so density curves of posterior distributions for other parameters are hardly visible.

stack
Logical, if TRUE, densities are plotted as stacked lines. Else, densities are plotted for each parameter among each other.

Details

data_plot() is in most situation not needed when the purpose is plotting, since most plot()-functions in see internally call data_plot() to prepare the data for plotting.

Many plot()-functions have a data-argument that is needed when the data or model for plotting can't be retrieved via data_plot(). In such cases, plot() gives an error and asks for providing data or models.

Most plot()-functions work out-of-the-box, i.e. you don't need to do much more than calling plot(<object>) (see 'Examples'). Some plot-functions allow to specify arguments to modify the transparancy or color of geoms, these are shown in the 'Usage' section.

Plot-functions are available for objects from following functions (note that functions from packages might be listed here that are currently still in development and probably not yet available):

bayestestR::bayesfactor_models()

bayestestR::bayesfactor_savagedickey()

bayestestR::ci()

bayestestR::equivalence_test()

bayestestR::estimate_density()

bayestestR::hdi()

bayestestR::p_direction()

bayestestR::rope()

estimate::estimateContrasts()

performance::binned_residuals()

performance::check_collinearity()

performance::check_distribution()

performance::check_heteroscedasticity()

performance::check_homogeneity()

performance::check_model()

performance::check_normality()

performance::check_outliers()

performance::performance_roc()

Examples

library(bayestestR)
library(rstanarm)

#> Loading required package: Rcpp

#> Registered S3 method overwritten by 'xts':
#>   method     from
#>   as.zoo.xts zoo 

#> rstanarm (Version 2.18.2, packaged: 2018-11-08 22:19:38 UTC)

#> - Do not expect the default priors to remain the same in future rstanarm versions.

#> Thus, R scripts should specify priors explicitly, even if they are just the defaults.

#> - For execution on a local, multicore CPU with excess RAM we recommend calling

#> options(mc.cores = parallel::detectCores())

#> - Plotting theme set to bayesplot::theme_default().

model <- stan_glm(
  Sepal.Length ~ Petal.Width * Species,
  data = iris,
  chains = 2, iter = 200
)

#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.001 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 10 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: WARNING: There aren't enough warmup iterations to fit the
#> Chain 1:          three stages of adaptation as currently configured.
#> Chain 1:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 1:          the given number of warmup iterations:
#> Chain 1:            init_buffer = 15
#> Chain 1:            adapt_window = 75
#> Chain 1:            term_buffer = 10
#> Chain 1: 
#> Chain 1: Iteration:   1 / 200 [  0%]  (Warmup)
#> Chain 1: Iteration:  20 / 200 [ 10%]  (Warmup)
#> Chain 1: Iteration:  40 / 200 [ 20%]  (Warmup)
#> Chain 1: Iteration:  60 / 200 [ 30%]  (Warmup)
#> Chain 1: Iteration:  80 / 200 [ 40%]  (Warmup)
#> Chain 1: Iteration: 100 / 200 [ 50%]  (Warmup)
#> Chain 1: Iteration: 101 / 200 [ 50%]  (Sampling)
#> Chain 1: Iteration: 120 / 200 [ 60%]  (Sampling)
#> Chain 1: Iteration: 140 / 200 [ 70%]  (Sampling)
#> Chain 1: Iteration: 160 / 200 [ 80%]  (Sampling)
#> Chain 1: Iteration: 180 / 200 [ 90%]  (Sampling)
#> Chain 1: Iteration: 200 / 200 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.084 seconds (Warm-up)
#> Chain 1:                0.096 seconds (Sampling)
#> Chain 1:                0.18 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'continuous' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 0 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: WARNING: There aren't enough warmup iterations to fit the
#> Chain 2:          three stages of adaptation as currently configured.
#> Chain 2:          Reducing each adaptation stage to 15%/75%/10% of
#> Chain 2:          the given number of warmup iterations:
#> Chain 2:            init_buffer = 15
#> Chain 2:            adapt_window = 75
#> Chain 2:            term_buffer = 10
#> Chain 2: 
#> Chain 2: Iteration:   1 / 200 [  0%]  (Warmup)
#> Chain 2: Iteration:  20 / 200 [ 10%]  (Warmup)
#> Chain 2: Iteration:  40 / 200 [ 20%]  (Warmup)
#> Chain 2: Iteration:  60 / 200 [ 30%]  (Warmup)
#> Chain 2: Iteration:  80 / 200 [ 40%]  (Warmup)
#> Chain 2: Iteration: 100 / 200 [ 50%]  (Warmup)
#> Chain 2: Iteration: 101 / 200 [ 50%]  (Sampling)
#> Chain 2: Iteration: 120 / 200 [ 60%]  (Sampling)
#> Chain 2: Iteration: 140 / 200 [ 70%]  (Sampling)
#> Chain 2: Iteration: 160 / 200 [ 80%]  (Sampling)
#> Chain 2: Iteration: 180 / 200 [ 90%]  (Sampling)
#> Chain 2: Iteration: 200 / 200 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.091 seconds (Warm-up)
#> Chain 2:                0.093 seconds (Sampling)
#> Chain 2:                0.184 seconds (Total)
#> Chain 2: 

x <- rope(model)

#> Possible multicollinearity between Petal.Width:Speciesversicolor and Petal.Width (r = 0.88), Petal.Width:Speciesversicolor and Speciesversicolor (r = 0.73), Petal.Width:Speciesvirginica and Petal.Width:Speciesversicolor (r = 0.82). This might lead to inappropriate results. See 'Details' in '?rope'.

plot(x)

#> Error: Failed at retrieving data :( Please provide original model or data through the `data` argument

x <- hdi(model)
plot(x) + theme_modern()

#> Error: Failed at retrieving data :( Please provide original model or data through the `data` argument

# NOT RUN {
data <- rnorm(1000, 1)
x <- p_direction(data)
plot(x)

x <- p_direction(model)
plot(x)

model <- stan_glm(
  mpg ~ wt + gear + cyl + disp,
  chains = 2,
  iter = 200,
  data = mtcars
)
x <- equivalence_test(model)
plot(x)
# }

# NOT RUN {
library(bayestestR)
library(see)

lm0 <- lm(qsec ~ 1, data = mtcars)
lm1 <- lm(qsec ~ drat, data = mtcars)
lm2 <- lm(qsec ~ wt, data = mtcars)
lm3 <- lm(qsec ~ drat + wt, data = mtcars)

result <- bayesfactor_models(lm1, lm2, lm3, denominator = lm0)

plot(result, n_pies = "one", value = "probability") + theme_modern() +
  scale_fill_pizza(reverse = TRUE)
# }