material_colors {see}	R Documentation

Extract material design colors as hex codes

Description

Can be used to get the hex code of specific colors from the material design color palette. Use material_colors() to see all available color.

Usage


material_colors(...)

Arguments

...
Character names of colors.

Value

A character vector with color-codes.

Examples


material_colors()

material_colors("indigo", "lime")

[Package see version 0.6.8 Index]