plot.see_check_normality {see}	R Documentation

Plot method for check model for (non-)normality of residuals

Description

The plot() method for the performance::check_normality() function.

Usage


## S3 method for class 'see_check_normality'
plot(
  x,
  type = c("density", "qq", "pp"),
  data = NULL,
  size_line = 0.8,
  size_point = 2,
  alpha = 0.2,
  dot_alpha = 0.8,
  colors = c("#3aaf85", "#1b6ca8"),
  detrend = FALSE,
  ...
)

Arguments

x
An object.

type
Character vector, indicating the type of plot.

data
The original data used to create this object. Can be a statistical model.

size_line
Numeric value specifying size of line geoms.

size_point
Numeric specifying size of point-geoms.

dot_alpha, alpha
Numeric value specifying alpha level of the confidence bands and point-geoms.

colors
Character vector of length two, indicating the colors (in hex-format) for points and line.

detrend
Logical that decides if the plot should be detrended.

...
Arguments passed to or from other methods.

Value

A ggplot2-object.

Examples


library(performance)
m <<- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
result <- check_normality(m)
plot(result)

[Package see version 0.6.8 Index]