plot.see_hdi {see}	R Documentation

Plot method for uncertainty or credible intervals

Description

The plot() method for the bayestestR::hdi() and related function.

Usage


## S3 method for class 'see_hdi'
plot(
  x,
  data = NULL,
  show_intercept = FALSE,
  show_zero = TRUE,
  show_title = TRUE,
  n_columns = 1,
  ...
)

Arguments

x
An object.

data
The original data used to create this object. Can be a statistical model.

show_intercept
Logical, if TRUE, the intercept-parameter is included in the plot. By default, it is hidden because in many cases the intercept-parameter has a posterior distribution on a very different location, so density curves of posterior distributions for other parameters are hardly visible.

show_zero
Logical. If TRUE, will add a vertical (dotted) line at 0.

show_title
Logical. If TRUE, will show the title of the plot.

n_columns
For models with multiple components (like fixed and random, count and zero-inflated), defines the number of columns for the panel-layout. If NULL, a single, integrated plot is shown.

...
Arguments passed to or from other methods.

Value

A ggplot2-object.

Examples



if (require("bayestestR") && require("rstanarm")) {
  set.seed(123)
  m <<- stan_glm(Sepal.Length ~ Petal.Width * Species, data = iris, refresh = 0)
  result <- hdi(m)
  result
  plot(result)
}


[Package see version 0.6.8 Index]