plot.see_parameters_distribution {see}	R Documentation

Plot method for describing distributions of vectors

Description

The plot() method for the parameters::describe_distribution() function.

Usage


## S3 method for class 'see_parameters_distribution'
plot(
  x,
  dispersion = FALSE,
  dispersion_alpha = 0.3,
  dispersion_color = "#3498db",
  dispersion_style = c("ribbon", "curve"),
  size_bar = 0.7,
  highlight = NULL,
  highlight_color = NULL,
  ...
)

Arguments

x
An object.

dispersion
Logical. If TRUE, a range of dispersion for each variable to the plot will be added.

dispersion_alpha
Numeric value specifying the transparency level of dispersion ribbon.

dispersion_color
Character specifying the color of dispersion ribbon.

dispersion_style
Character describing the style of dispersion area. "ribbon" for a ribbon, "curve" for a normal-curve.

size_bar
Size of bar geoms.

highlight
A vector with names of categories in x that should be highlighted.

highlight_color
A vector of color values for highlighted categories. The remaining (non-highlighted) categories will be filled with a lighter grey.

...
Arguments passed to or from other methods.

Value

A ggplot2-object.

Examples


library(parameters)
set.seed(333)
x <- sample(1:100, 1000, replace = TRUE)
result <- describe_distribution(x)
result
plot(result)

[Package see version 0.6.8 Index]