Plot method for Bayes Factors for a single parameter

The plot() method for the bayestestR::bayesfactor_parameters() function.

Usage

# S3 method for class 'see_bayesfactor_parameters'
plot(
  x,
  size_point = 2,
  rope_color = "#0171D3",
  rope_alpha = 0.2,
  show_intercept = FALSE,
  ...
)

Arguments

x

An object.

size_point

Numeric specifying size of point-geoms.

rope_color

Character specifying color of ROPE ribbon.

rope_alpha

Numeric specifying transparency level of ROPE ribbon.

show_intercept

Logical, if TRUE, the intercept-parameter is included in the plot. By default, it is hidden because in many cases the intercept-parameter has a posterior distribution on a very different location, so density curves of posterior distributions for other parameters are hardly visible.

...

Arguments passed to or from other methods.

Value

A ggplot2-object.